Code: Remove code duplication
