+++ jarhc/src/main/java/org/jarhc/artifacts/DepsDevApiVulnerabilityFinder.java
@@ -16,9 +16,6 @@
 
 package org.jarhc.artifacts;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.net.URLEncoder;
@@ -28,8 +25,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicBoolean;
-import org.jarhc.utils.IOUtils;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -50,7 +45,7 @@ import org.slf4j.LoggerFactory;
  * of this finder. (Durable, time-limited caching is provided by the surrounding
  * cache decorators.)
  */
-public class DepsDevAPIVulnerabilityFinder implements VulnerabilityFinder {
+public class DepsDevApiVulnerabilityFinder implements VulnerabilityFinder {
 
 	// Example request URLs:
 	// https://api.deps.dev/v3/systems/MAVEN/packages/org.apache.logging.log4j%3Alog4j-core/versions/2.14.1
@@ -62,24 +57,24 @@ public class DepsDevAPIVulnerabilityFinder implements VulnerabilityFinder {
 
 	private final Logger logger;
 	private final DepsDevSettings settings;
-
-	private final AtomicBoolean statusInfoLogged = new AtomicBoolean(false);
+	private final DepsDevApiClient client;
 
 	// in-memory cache for advisory details, de-duplicating shared advisories
 	// across multiple artifacts within a single run
 	private final Map<String, Vulnerability> advisoryCache = new ConcurrentHashMap<>();
 
-	public DepsDevAPIVulnerabilityFinder() {
-		this(LoggerFactory.getLogger(DepsDevAPIVulnerabilityFinder.class), DepsDevSettings.fromSystemProperties());
+	public DepsDevApiVulnerabilityFinder() {
+		this(LoggerFactory.getLogger(DepsDevApiVulnerabilityFinder.class), DepsDevSettings.fromSystemProperties());
 	}
 
-	public DepsDevAPIVulnerabilityFinder(Logger logger) {
+	public DepsDevApiVulnerabilityFinder(Logger logger) {
 		this(logger, DepsDevSettings.fromSystemProperties());
 	}
 
-	DepsDevAPIVulnerabilityFinder(Logger logger, DepsDevSettings settings) {
+	DepsDevApiVulnerabilityFinder(Logger logger, DepsDevSettings settings) {
 		this.logger = logger;
 		this.settings = settings;
+		this.client = new DepsDevApiClient(logger, settings);
 	}
 
 	@Override
@@ -107,7 +102,7 @@ public class DepsDevAPIVulnerabilityFinder implements VulnerabilityFinder {
 		String version = artifact.getVersion();
 		URL url = createUrl(settings.getBaseUrl() + VERSION_PATH, encode(name), encode(version));
 
-		Optional<String> text = downloadText(url);
+		Optional<String> text = client.downloadText(url);
 		if (text.isEmpty()) {
 			// version not found (API responded with HTTP 404)
 			if (logger.isDebugEnabled()) {
@@ -139,7 +134,7 @@ public class DepsDevAPIVulnerabilityFinder implements VulnerabilityFinder {
 		}
 
 		URL url = createUrl(settings.getBaseUrl() + ADVISORY_PATH, encode(advisoryId));
-		Optional<String> text = downloadText(url);
+		Optional<String> text = client.downloadText(url);
 		if (text.isEmpty()) {
 			// advisory not found (API responded with HTTP 404)
 			logger.warn("Advisory not found: {}", advisoryId);
@@ -198,68 +193,6 @@ public class DepsDevAPIVulnerabilityFinder implements VulnerabilityFinder {
 		} catch (JSONException e) {
 			throw new RepositoryException("JSON parser error for URL: " + url, e);
 		}
-	}
-
-	/**
-	 * Download the response body for the given URL.
-	 *
-	 * @param url URL
-	 * @return Response body, or an empty {@link Optional} if the API responded with HTTP 404.
-	 * @throws RepositoryException If an unexpected I/O error occurs.
-	 */
-	private Optional<String> downloadText(URL url) throws RepositoryException {
-		try {
-			Optional<byte[]> data = downloadFile(url);
-			return data.map(bytes -> new String(bytes, StandardCharsets.UTF_8));
-		} catch (IOException e) {
-			// log information about deps.dev status once
-			if (statusInfoLogged.compareAndSet(false, true)) {
-				logger.warn("Problem with api.deps.dev. Visit https://deps.dev to check the status of the deps.dev API.");
-			}
-			throw new RepositoryException("Unexpected I/O error for URL: " + url, e);
-		}
-	}
-
-	private Optional<byte[]> downloadFile(URL url) throws IOException {
-
-		HttpURLConnection connection = null;
-		try {
-
-			// get connection settings
-			int timeout = settings.getTimeout() * 1000;
-			Map<String, String> headers = settings.getHeaders();
-
-			// prepare connection
-			connection = (HttpURLConnection) url.openConnection();
-			connection.setRequestMethod("GET");
-			for (Map.Entry<String, String> header : headers.entrySet()) {
-				connection.setRequestProperty(header.getKey(), header.getValue());
-			}
-			connection.setConnectTimeout(timeout);
-			connection.setReadTimeout(timeout);
-			connection.setDoOutput(false);
-			connection.setDoInput(true);
-
-			// get response
-			int status = connection.getResponseCode();
-			if (status == 404) {
-				// not found
-				return Optional.empty();
-			} else if (status != 200) {
-				throw new IOException("Unexpected status code '" + status + "' for URL: " + url);
-			}
-
-			try (InputStream stream = connection.getInputStream()) {
-				byte[] data = IOUtils.toByteArray(stream);
-				return Optional.of(data);
-			}
-
-		} finally {
-			if (connection != null) {
-				connection.disconnect();
-			}
-		}
-
 	}
 
 }
