+++ jarhc/src/test/java/org/jarhc/artifacts/DepsDevApiVulnerabilityFinderTest.java
@@ -30,11 +30,13 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.slf4j.Logger;
 
-class DepsDevAPIArtifactFinderTest {
+class DepsDevApiVulnerabilityFinderTest {
 
-	private final Logger logger = LoggerBuilder.collect(DepsDevAPIArtifactFinder.class);
+	private final Logger logger = LoggerBuilder.collect(DepsDevApiVulnerabilityFinder.class);
 	private final DepsDevSettings settings = Mockito.mock(DepsDevSettings.class);
-	private final DepsDevAPIArtifactFinder artifactFinder = new DepsDevAPIArtifactFinder(logger, settings);
+	private final DepsDevApiVulnerabilityFinder finder = new DepsDevApiVulnerabilityFinder(logger, settings);
+
+	private final Artifact artifact = new Artifact("org.example", "lib", "1.0", "jar");
 
 	@BeforeEach
 	void setUp() {
@@ -44,57 +46,36 @@ class DepsDevAPIArtifactFinderTest {
 	}
 
 	@Test
-	void test_findArtifact_withInvalidChecksum() {
-
-		// test
-		assertThrows(IllegalArgumentException.class, () -> artifactFinder.findArtifacts("this is NOT a checksum!"));
-
-		// assert
+	void findVulnerabilities_rejectsNullArtifact() {
+		assertThrows(IllegalArgumentException.class, () -> finder.findVulnerabilities(null));
 		assertLogger(logger).isEmpty();
 	}
 
 	@Test
-	void test_findArtifact_withInvalidURL() {
+	void findVulnerabilities_withMalformedUrl() {
 
-		// prepare
 		doReturn("htp://localhost").when(settings).getBaseUrl();
 
-		// test
-		Exception result = assertThrows(RepositoryException.class, () -> artifactFinder.findArtifacts("1234"));
+		Exception result = assertThrows(RepositoryException.class, () -> finder.findVulnerabilities(artifact));
 
-		// assert
 		assertThat(result)
-				.hasMessage("Malformed URL for checksum: 1234")
-				.hasCause(new MalformedURLException("unknown protocol: htp"));
+				.hasMessageStartingWith("Malformed URL: htp://localhost/systems/MAVEN/packages/")
+				.hasCauseInstanceOf(MalformedURLException.class);
 		assertLogger(logger).isEmpty();
 	}
 
 	@Test
-	void test_findArtifact_withUnknownURL() {
+	void findVulnerabilities_withUnknownHost() {
 
-		// prepare
 		doReturn("http://unknown").when(settings).getBaseUrl();
 
-		// test
-		Exception result = assertThrows(RepositoryException.class, () -> artifactFinder.findArtifacts("1234"));
+		Exception result = assertThrows(RepositoryException.class, () -> finder.findVulnerabilities(artifact));
 
-		// assert
 		assertThat(result)
-				.hasMessageStartingWith("Unexpected I/O error for URL: http://unknown/query?")
-				.hasCause(new UnknownHostException("unknown"));
+				.hasMessageStartingWith("Unexpected I/O error for URL: http://unknown/systems/MAVEN/packages/")
+				.hasCauseInstanceOf(UnknownHostException.class);
 		assertLogger(logger)
 				.hasWarn("Problem with api.deps.dev. Visit https://deps.dev to check the status of the deps.dev API.")
-				.isEmpty();
-
-		// test
-		result = assertThrows(RepositoryException.class, () -> artifactFinder.findArtifacts("1234"));
-
-		// assert
-		assertThat(result)
-				.hasMessageStartingWith("Unexpected I/O error for URL: http://unknown/query?")
-				.hasCause(new UnknownHostException("unknown"));
-		assertLogger(logger)
-				// no additional warning
 				.isEmpty();
 	}
