+++ jarhc/src/test/java/org/jarhc/it/AbstractTest.java
@@ -31,7 +31,7 @@ import org.jarhc.app.ArtifactSource;
 import org.jarhc.app.JarSource;
 import org.jarhc.app.Options;
 import org.jarhc.artifacts.ArtifactFinder;
-import org.jarhc.artifacts.DepsDevAPIVulnerabilityFinder;
+import org.jarhc.artifacts.DepsDevApiVulnerabilityFinder;
 import org.jarhc.artifacts.MavenRepository;
 import org.jarhc.artifacts.Repository;
 import org.jarhc.artifacts.VulnerabilityFinder;
@@ -80,7 +80,7 @@ abstract class AbstractTest {
 		}
 
 		// vulnerability finder backed by the mock server (URLs set by MavenProxyServerExtension)
-		VulnerabilityFinder vulnerabilityFinder = new DepsDevAPIVulnerabilityFinder();
+		VulnerabilityFinder vulnerabilityFinder = new DepsDevApiVulnerabilityFinder();
 
 		// prepare an injector
 		Injector injector = new Injector();
