+++ jarhc/src/test/java/org/jarhc/it/ApplicationTest.java
@@ -29,7 +29,7 @@ import org.jarhc.TestUtils;
 import org.jarhc.app.Application;
 import org.jarhc.app.Options;
 import org.jarhc.artifacts.ArtifactFinder;
-import org.jarhc.artifacts.DepsDevAPIVulnerabilityFinder;
+import org.jarhc.artifacts.DepsDevApiVulnerabilityFinder;
 import org.jarhc.artifacts.MavenRepository;
 import org.jarhc.it.utils.ArtifactFinderMock;
 import org.jarhc.it.utils.MavenProxyServerExtension;
@@ -67,7 +67,7 @@ class ApplicationTest {
 		application.setRepository(repository);
 
 		// vulnerability finder backed by the mock server (URLs set by MavenProxyServerExtension)
-		application.setVulnerabilityFinder(new DepsDevAPIVulnerabilityFinder());
+		application.setVulnerabilityFinder(new DepsDevApiVulnerabilityFinder());
 	}
 
 	@Test
